refactor(dftu): split DFT+U force/stress into r-space NAO files and rename for clarity
Rename and reorganize DFT+U force/stress implementation:

New files (r-space, NAO basis):
- dftu_nao_fs_r.h/cpp    : unified force+stress entry with full formula docs
- dftu_nao_for_r.h/cpp   : single-pair force core (cal_for_IJR_nao_r)
- dftu_nao_str_r.h/cpp   : single-pair stress core (cal_str_IJR_nao_r)

Renamed files:
- dftu_lcao_op.h/cpp          -> dftu_nao_op.h/cpp
- dftu_lcao_op_legacy.cpp     -> dftu_nao_op_legacy.cpp
- dftu_force.h/cpp (legacy k) -> dftu_nao_fs_k.h/cpp
- dftu_folding.h/cpp          -> dftu_nao_folding.h/cpp

Deleted:
- dftu_fs.cpp (split into dftu_nao_fs_r.cpp + dftu_nao_for_r.cpp + dftu_nao_str_r.cpp)

Updated:
- dftu_nao_op.h: expose getters and cal_pot_onsite/transfer_pot_onsite for free functions
- force_stress_lcao.cpp, hamilt_lcao.cpp, dftu_lcao_occ.cpp: update includes
- test/CMakeLists.txt, test/dftu_lcao_test.cpp: update source list and include

Naming convention:
- nao : NAO/LCAO basis set
- _r  : real-space implementation (uses DMR, k-point independent)
- _k  : k-space legacy implementation (deprecated, broken)
- for : force
- str : stress
- fs  : force+stress combined
