+++ source/source_lcao/module_dftu/CMakeLists.txt
@@ -1,7 +1,7 @@
 list(APPEND objects
       dftu_lcao.cpp
-      dftu_force.cpp
-      dftu_folding.cpp
+      dftu_nao_fs_k.cpp
+      dftu_nao_folding.cpp
       dftu_lcao_pots.cpp
       dftu_lcao_occ.cpp
       dftu_lcao_energy.cpp
@@ -10,9 +10,11 @@ list(APPEND objects
 
 if(ENABLE_LCAO)
   list(APPEND objects
-        dftu_lcao_op.cpp
-        dftu_fs.cpp
-        dftu_lcao_op_legacy.cpp
+        dftu_nao_op.cpp
+        dftu_nao_fs_r.cpp
+        dftu_nao_for_r.cpp
+        dftu_nao_str_r.cpp
+        dftu_nao_op_legacy.cpp
   )
 endif()
