+++ source/source_lcao/module_dftu/dftu_nao_fs_k.cpp
@@ -1,6 +1,6 @@
 #ifdef __LCAO
-#include "dftu_force.h"
-#include "dftu_folding.h"
+#include "dftu_nao_fs_k.h"
+#include "dftu_nao_folding.h"
 #include "dftu_lcao.h"
 #include "dftu_lcao_pots.h"
 #include "source_base/global_function.h"
